+++ tailwind.config.cjs
@@ -1,19 +1,11 @@
 const plugin = require('tailwindcss/plugin');
 
-const spacing = 0.25;
-const unit = 'rem';
-const space = time => `${time * spacing}${unit}`;
 
 
 module.exports = {
 	content: ['./src/index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
 	theme: {
 		extend: {
-			spacing: {
-				42: space(42),
-				78: space(78),
-				80: space(80),
-			},
 			boxShadow: {
 				mdd: '0 4px 6px -1px rgba(0, 0, 0, 0.2), 0 2px 4px -1px rgba(0, 0, 0, 0.1)',
 			},
@@ -28,12 +20,12 @@ module.exports = {
 	plugins: [
 		plugin(({ addUtilities, matchUtilities, theme }) => {
 			addUtilities({
-				// inblock=inline-block + 顶部对齐
+				// inblock = inline-block + align-top
 				'.inblock': {
 					display: 'inline-block',
 					verticalAlign: 'top'
 				},
-				// 文本溢出省略号
+				// text ellipsis
 				'.elli': {
 					overflow: 'hidden',
 					whiteSpace: 'nowrap',
@@ -41,7 +33,7 @@ module.exports = {
 				},
 			});
 
-			// 动画延迟
+			// transition delay
 			matchUtilities(
 				{
 					trans: duration => ({