+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## v1.1.0 - 2022.08.23 11
+* use `vite` lib mode to build code, instead export directly
+* update `jsconfig.json`
+* update `package.json`.`dependencies` and use `peerDependencies`
+* sort `package.json` keys
+* bump up versions
+* simplify `tailwind.config.cjs ` and `vite.config.js`
+
+
+
 ## v1.0.0 - 2022.08.22 17
 * independent from many projects
 * tweak all files for publishing to npm